+++ src/providers/twitch/api/Helix.hpp
@@ -134,6 +134,27 @@ struct HelixGame {
     }
 };
 
+struct HelixTag {
+    QString id;
+    bool isAuto;
+    QString englishName;
+    QString englishDescription;
+    // TODO: If chatterino gets localization, make this a map<qstring, qstring>
+    explicit HelixTag(QJsonObject jsonObject)
+        : id(jsonObject.value("tag_id").toString())
+        , isAuto(jsonObject.value("is_auto").toBool())
+        , englishName(jsonObject.value("localization_names")
+                          .toObject()
+                          .value("en-us")
+                          .toString())
+        , englishDescription(jsonObject.value("localization_descriptions")
+                                 .toObject()
+                                 .value("en-us")
+                                 .toString())
+    {
+    }
+};
+
 class Helix final : boost::noncopyable
 {
 public:
@@ -183,6 +204,21 @@ public:
     void searchGames(QString query,
                      ResultCallback<std::vector<HelixGame>> successCallback,
                      HelixFailureCallback failureCallback);
+
+    void updateStreamTags(QString broadcasterId, QStringList tags,
+                          std::function<void()> successCallback,
+                          HelixFailureCallback failureCallback);
+
+    void getStreamTags(QString broadcasterId,
+                       ResultCallback<std::vector<HelixTag>> successCallback,
+                       HelixFailureCallback failureCallback);
+
+    void getAllStreamTags(ResultCallback<std::vector<HelixTag>> successCallback,
+                          HelixFailureCallback failureCallback);
+    void fetchStreamTags(
+        QString after,
+        ResultCallback<std::vector<HelixTag>, QString> successCallback,
+        HelixFailureCallback failureCallback);
 
     void getGameById(QString gameId, ResultCallback<HelixGame> successCallback,
                      HelixFailureCallback failureCallback);