+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Major: Add /settitle (Mm2PL/chatterino2#8)
 - Major: Add /setgame (Mm2PL/chatterino2#9)
 - Major: Add GUI for /settitle, /setgame (Mm2PL/chatterino#11)
+- Major: Edit stream tags in GUI (Mm2PL/chatterino#12)
 
 ### Chatterino